Fix multi-ore prospecting (closes #219)
The issue addressed in #219 was that Granite would trigger the ProPick to detect "Autunite and Granite"

Upon testing, a single granite would not do this, but more than 1 granite block would. It occured to me that this was because the ArrayList impl allowed duplicates.

The fix to this issue was to use a HashSet instead, which does not allow duplicates.